moved all plotting functions from utils to utilsplot
