+++ Example_02_04.py
@@ -1,7 +1,7 @@
 import matplotlib.pyplot as plt
 import numpy as np
-from utils import feedback, tf, bodeclosedloop, marginsclosedloop, ControllerTuning
-from utilsplot import step_response_plot
+from utils import feedback, tf, marginsclosedloop, ControllerTuning
+from utilsplot import step_response_plot, bodeclosedloop
 
 s = tf([1, 0], 1)
 G = 3 * (-2 * s + 1) / ((10 * s + 1) * (5 * s + 1))
