+++ Example_02_05.py
@@ -1,7 +1,7 @@
 import numpy as np
 import matplotlib.pyplot as plt
-from utils import feedback, bodeclosedloop, tf, marginsclosedloop
-from utilsplot import step_response_plot
+from utils import feedback, tf, marginsclosedloop
+from utilsplot import step_response_plot, bodeclosedloop
 
 s = tf([1, 0], 1)
 G = 4 /((s - 1) * (0.02 * s + 1)**2)
@@ -21,8 +21,8 @@ bodeclosedloop(G, K, -1, 2, label='Figure 2.15', margin=True)
 #TODO there is a descrepancy with the phase plots
 
 GM, PM, wc, wb, wbt, valid = marginsclosedloop(L) 
-print 'GM :' , GM
-print "PM:", np.round(PM*180/np.pi, 1), "deg or", np.round(PM, 2), "rad"
-print 'wb :' , wb
-print 'wc :' , wc
-print 'wbt :' , wbt
+print 'GM:' , np.round(GM, 2)
+print 'PM:', np.round(PM / 180 * np.pi, 2), "rad or", np.round(PM, 2), "deg"
+print 'wb:' , np.round(wb, 2)
+print 'wc:' , np.round(wc, 2)
+print 'wbt:' , np.round(wbt, 4)
