+++ frequency_response.py
@@ -1,20 +1,11 @@
-import numpy
 import matplotlib.pyplot as plt
+from utilsplot import bode
+from utils import tf
 
-def unwrapped_angle(Gfr):
-    return numpy.unwrap(numpy.angle(Gfr))
-
-def bode(w, Gfr):
-    plt.subplot(2, 1, 1)
-    plt.loglog(w, numpy.abs(Gfr))
-    plt.subplot(2, 1, 2)
-    plt.semilogx(w, unwrapped_angle(Gfr))
-
-w = numpy.logspace(-2, 2)
-s = w * 1j
+s = tf([1, 0])
 
 for power in range(1, 4):
-    Gfr = 1/(s + 1)**power
-    bode(w, Gfr)
+    G = 1/(s + 1)**power
+    bode(G)
 
 plt.show()
