+++ utilsplot.py
@@ -24,11 +24,15 @@ show : boolean
 
 Plots
 -----
-bode_plot :
-    (Can bring across from utils.py)
-    
-nyquist_plot:
-    (Can bring across from utils.py)
+bode: Shows the bode plot for a plant model
+
+bodeclosedloop: Shows the bode plot for a controller model
+
+mimo_bode: Plots the max and min singular values of G and computes the crossover freq
+
+nyquist_plot: TODO
+
+mino_nyquist_plot: Nyquist stability plot for MIMO system
     
 sv_plot : Maximum and minimum singular values of a matirix
     
@@ -37,12 +41,6 @@ condtn_nm_plot : A plot of the condition number for a specified diagonal
 rga_plot: A plot of the relative gain interactions for a matrix over a given frequency
 
 rga_nm_plot: A plot of the RGA number for a given pairing
-
-weightd_sens_plot :
-    (Can bring across from utils.py)
-    
-perf_wt_plot :
-    (Can bring across from utils.py)
     
 dis_rejctn_plot : A plot of the disturbance condition number and the bounds imposed
     by the singular values.
@@ -51,7 +49,9 @@ freq_step_response_plot: A subplot for both the frequnecy response and step
     response for a controlled plant
     
 step_response_plot: A plot of the step response of a transfer function
-    
+   
+perf_Wp_plot: MIMO sensitivity S and performance weight Wp plotting funtion
+   
 """
 
 import numpy #do not abbreviate this module as np in utilsplot.py
@@ -59,6 +59,306 @@ import utils
 import matplotlib.pyplot as plt
 import sys
       
+
+def bode(G, w_start=-2, w_end=2, axlim=None, points=100, margin=False):
+    """ 
+    Shows the bode plot for a plant model
+    
+    Parameters
+    ----------
+    G : tf
+        plant transfer function
+    margin : boolean
+        show the cross over frequencies on the plot (optional)        
+          
+    Returns
+    -------
+    GM : array containing a real number      
+        gain margin
+    PM : array containing a real number           
+        phase margin         
+    """
+
+    GM, PM, wc, w_180 = utils.margins(G)
+
+    # plotting of Bode plot and with corresponding frequencies for PM and GM
+#    if ((w2 < numpy.log(w_180)) and margin):
+#        w2 = numpy.log(w_180)  
+    w = numpy.logspace(w_start, w_end, points)
+    s = 1j*w
+
+    plt.figure('Bode Plot')
+    # Magnitude of G(jw)
+    plt.subplot(211)
+    gains = numpy.abs(G(s))
+    plt.loglog(w, gains)
+    if margin:
+        plt.axvline(w_180, color='black')
+        plt.text(w_180, numpy.average([numpy.max(gains), numpy.min(gains)]), r'$\angle$G(jw) = -180$\degree$')
+    plt.axhline(1., color='red')
+    plt.grid()
+    plt.ylabel('Magnitude')
+
+    # Phase of G(jw)
+    plt.subplot(212)
+    phaseangle = utils.phase(G(s), deg=True)
+    plt.semilogx(w, phaseangle)
+    if margin:
+        plt.axvline(wc, color='black')
+        plt.text(wc, numpy.average([numpy.max(phaseangle), numpy.min(phaseangle)]), '|G(jw)| = 1')
+    plt.axhline(-180., color='red')
+    plt.grid()
+    plt.ylabel('Phase')
+    plt.xlabel('Frequency [rad/unit time]')
+    
+
+    return GM, PM
+    
+def bodeclosedloop(G, K, w1, w2, label='Figure', margin=False):
+    """ 
+    Shows the bode plot for a controller model
+    
+    Parameters
+    ----------
+    G : tf
+        plant transfer function
+    K : tf
+        controller transfer function
+    w1 : real
+        start frequency
+    w2 : real
+        end frequency
+    label : string
+        title for the figure (optional)
+    margin : boolean
+        show the cross over frequencies on the plot (optional)             
+    """
+    
+    w = numpy.logspace(w1, w2, 1000)    
+    L = G(1j*w) * K(1j*w)
+    S = utils.feedback(1, L)
+    T = utils.feedback(L, 1)
+    
+    plt.figure(label)
+    plt.subplot(2, 1, 1)
+    plt.loglog(w, abs(L))
+    plt.loglog(w, abs(S))
+    plt.loglog(w, abs(T))
+    plt.grid()
+    plt.ylabel("Magnitude")
+    plt.legend(["L", "S", "T"],
+               bbox_to_anchor=(0, 1.01, 1, 0), loc=3, ncol=3)
+    
+    if margin:        
+        plt.plot(w, 1/numpy.sqrt(2) * numpy.ones(len(w)), linestyle='dotted')
+        
+    plt.subplot(2, 1, 2)
+    plt.semilogx(w, utils.phase(L, deg=True))
+    plt.semilogx(w, utils.phase(S, deg=True))
+    plt.semilogx(w, utils.phase(T, deg=True))
+    plt.grid()
+    plt.ylabel("Phase")
+    plt.xlabel("Frequency [rad/s]")  
+    
+
+def mimoBode(Gin, wStart, wEnd, Kin=None): 
+    """
+    Plots the max and min singular values of G and computes the crossover freq.
+    
+    If a controller is specified, the max and min singular values
+    of S are also plotted and the bandwidth freq computed.
+              
+    Parameters
+    ----------
+    Gin : numpy array
+        Matrix of plant transfer functions.
+    
+    wStart : float
+        Minimum power of w for the frequency range in rad/time. 
+        eg: for w startig at 10e-3, wStart = -3.
+        
+    wEnd : float
+        Maximum value of w for the frequency range in rad/time. 
+        eg: for w ending at 10e3, wStart = 3.
+    
+    Kin : numpy array
+        Controller matrix (optional).
+    
+    Returns
+    -------
+    wC : real
+        Crossover frequency.
+        
+    wB : real
+        Bandwidth frequency.
+        
+    Plot : matplotlib plot
+        Bode plot of singular values of G and S(optional).
+    
+    Example
+    -------
+    >>> K = numpy.array([[1., 2.],
+    ...                  [3., 4.]])*10
+    >>> t1 = numpy.array([[5., 5.],
+    ...                   [5., 5.]])
+    >>> t2 = numpy.array([[5., 6.],
+    ...                   [7., 8.]])
+    >>>                   
+    >>> def G(s):
+    ...     return(K*numpy.exp(-t1*s)/(t2*s + 1.))
+    >>>
+    >>> def Kc(s):
+    ...     return(numpy.array([[0.1, 0.],
+    ...                         [0., 0.1]])*10.)
+    >>> mimoBode(G, -3, 3, Kc)
+    Bandwidth is a tuple of wC, wB
+    (0.55557762223988783, 1.3650078065460138)
+    
+    """
+    xmin = 10**wStart
+    xmax = 10**wEnd
+    w = numpy.logspace(wStart, wEnd, 1000)
+    s = w*1j
+    Sv1 = numpy.zeros(len(w), dtype=complex)
+    Sv2 = numpy.zeros(len(w), dtype=complex)
+    f = 0
+    wC = 0
+    for i in range(len(w)):
+        Sv1[i] = utils.sigmas(Gin(s[i]))[0]
+        Sv2[i] = utils.sigmas(Gin(s[i]))[-1]
+        if (f < 1 and Sv2[i] < 1):
+            wC = w[i]
+            f = 1
+    ymin = numpy.min(Sv2)
+    plt.figure('MIMO Bode')
+    plt.clf()
+    plt.loglog(w, Sv1, 'k-', label='Max $\sigma$(G)')
+    plt.loglog(w, Sv2, 'k-', alpha=0.5, label='Min $\sigma$(G)')
+    plt.axhline(1, ls=':', lw=2, color='blue')
+    plt.text(xmin, 1.1, 'Mag = 1', color='blue')
+    plt.axvline(wC, ls=':', lw=2, color='blue')
+    plt.text(wC*1.1, ymin*1.1, 'wC', color='blue')
+    plt.legend(loc='upper right', fontsize = 10, ncol=1)
+    plt.xlabel('Frequency [rad/time]')
+    plt.ylabel('Magnitude')
+    plt.axis([xmin, xmax, None, None])
+    plt.grid(True)
+    fig = plt.gcf()
+    BG = fig.patch
+    BG.set_facecolor('white')
+    
+    if Kin is None:
+        Bandwidth = wC
+        print('Bandwidth = wC')
+    else:
+        def S(s):
+            L = Kin(s)*Gin(s)
+            dim1, dim2 = numpy.shape(Gin(0))
+            return(numpy.linalg.inv(numpy.eye(dim1) + L))      #SVD of S = 1/(I + L)
+        w = numpy.logspace(wStart, wEnd, 1000)
+        s = w*1j
+        Sv1 = numpy.zeros(len(w), dtype=complex)
+        Sv2 = numpy.zeros(len(w), dtype=complex)
+        f = 0
+        wB = 0
+        for i in range(len(w)):
+            Sv1[i] = utils.sigmas(S(s[i]))[0]
+            Sv2[i] = utils.sigmas(S(s[i]))[-1]
+            if (f < 1 and Sv1[i] > 0.707):
+                wB = w[i]
+                f = 1
+        plt.figure()
+        plt.loglog(w, Sv1, 'r-', label='Max $\sigma$(S)')
+        plt.loglog(w, Sv2, 'r-', alpha=0.5, label='Min $\sigma$(S)')
+        plt.axhline(0.707, ls=':', lw=2, color='green')
+        plt.text(xmin, 0.5, 'Mag = 0.707', color='green')
+        plt.axvline(wB, ls=':', lw=2, color='green')
+        plt.text(wB*1.1, ymin*1.1, 'wB', color='green')
+        plt.legend(loc='upper right', fontsize = 10, ncol=1)
+        plt.xlabel('Frequency [rad/time]')
+        plt.ylabel('Magnitude')
+        plt.axis([xmin, xmax, None, None])
+        plt.grid(True)
+        fig = plt.gcf()
+        BG = fig.patch
+        BG.set_facecolor('white')
+        Bandwidth = wC, wB
+        print('Bandwidth is a tuple of wC, wB')
+    return(Bandwidth)
+
+
+def mino_nyquist_plot(L, axLim, wStart, wEnd):
+    """
+    Nyquist stability plot for MIMO system.
+    
+    Parameters
+    ----------
+    L : numpy array
+        Closed loop transfer function matrix as a function of s, i.e. def L(s).
+    
+    axLim : float
+        Axis limit for square axis.  axis will run from -axLim to +axLim.
+    
+    wStart : float
+        Minimum power of w for the frequency range in rad/time. 
+        eg: for w startig at 10e-3, wStart = -3.
+        
+    wEnd : float
+        Maximum value of w for the frequency range in rad/time. 
+        eg: for w ending at 10e3, wStart = 3.
+        
+    Returns
+    -------
+    Nyquist stability plot.
+    
+    Example
+    -------
+    >>> K = numpy.array([[1., 2.],
+    ...                  [3., 4.]])
+    >>> t1 = numpy.array([[5., 5.],
+    ...                   [5., 5.]])
+    >>> t2 = numpy.array([[5., 6.],
+    ...                   [7., 8.]]) 
+    >>> Kc = numpy.array([[0.1, 0.], 
+    ...                   [0., 0.1]])*6
+    >>> 
+    >>> def G(s):
+    ...     return(K*numpy.exp(-t1*s)/(t2*s + 1))
+    ... 
+    >>> def L(s):
+    ...     return(Kc*G(s))
+    ... 
+    >>> #MIMOnyqPlot(L, 2)
+    
+    """
+    w = numpy.logspace(wStart, wEnd, 1000)    
+    Lin = numpy.zeros((len(w)), dtype=complex)
+    x = numpy.zeros((len(w)))
+    y = numpy.zeros((len(w)))
+    dim = numpy.shape(L(0.1))
+    for i in range(len(w)):        
+        Lin[i] = numpy.linalg.det(numpy.eye(dim[0]) + L(w[i]*1j))
+        x[i] = numpy.real(Lin[i])
+        y[i] = numpy.imag(Lin[i])        
+    plt.figure('MIMO Nyquist Plot')
+    plt.clf()
+    plt.plot(x, y, 'k-', lw=1)
+    plt.xlabel('Re G(wj)')
+    plt.ylabel('Im G(wj)')
+    # plotting a unit circle
+    x = numpy.linspace(-1, 1, 200)
+    y_up = numpy.sqrt(1-(x)**2)
+    y_down = -1*numpy.sqrt(1 - (x)**2)
+    plt.plot(x, y_up, 'b:', x, y_down, 'b:', lw=2)
+    plt.plot(0, 0, 'r*', ms=10)
+    plt.grid(True)
+    n = axLim           # Sets x-axis limits
+    plt.axis('equal')   # Ensure the unit circle remains round on resizing the figure
+    plt.axis([-n, n, -n, n])
+    fig = plt.gcf()
+    BG = fig.patch
+    BG.set_facecolor('white')
+
 
 def sv_plot(G, w_start=-2, w_end=2, axlim=None, points=100):
     '''
@@ -576,5 +876,109 @@ def step_response_plot(Y, U, t_end=50, initial_val=0, timedim='sec', axlim=None,
     plt.plot([0, t_end], numpy.ones(2),'--')    
     
     plt.axis(axlim)
-    plt.xlabel('Time [' + timedim + ']')
+    plt.xlabel('Time [' + timedim + ']')  
+
+
+def perf_Wp_plot(S, wB_req, maxSSerror, wStart, wEnd):
+    """
+    MIMO sensitivity S and performance weight Wp plotting funtion.
+    
+    Parameters
+    ----------
+    S : numpy array
+        Sensitivity transfer function matrix as function of s => S(s)
+        
+    wB_req : float
+        The design or require bandwidth of the plant in rad/time.
+        1/time eg: wB_req = 1/20sec = 0.05rad/s
+        
+    maxSSerror : float
+        The maximum stead state tracking error required of the plant.
+        
+    wStart : float
+        Minimum power of w for the frequency range in rad/time. 
+        eg: for w startig at 10e-3, wStart = -3.
+        
+    wEnd : float
+        Maximum value of w for the frequency range in rad/time. 
+        eg: for w ending at 10e3, wStart = 3.
+
+    Returns
+    -------
+    wB : float
+        The actualy plant bandwidth in rad/time given the specified controller 
+        used to generate the sensitivity matrix S(s).
+    
+    Plot : matplotlib figure
+        A plot of the sensitivity function and the performance weight across the
+        frequency range specified.
+        
+    Example
+    -------
+    >>> K = numpy.array([[1., 2.],
+    ...                  [3., 4.]])
+    >>> t1 = numpy.array([[5., 5.],
+    ...                   [5., 5.]])
+    >>> t2 = numpy.array([[5., 6.],
+    ...                   [7., 8.]])
+    >>> Kc = numpy.array([[0.1, 0.],
+    ...                   [0., 0.1]])*10
+    >>> 
+    >>> def G(s):
+    ...     return(K*numpy.exp(-t1*s)/(t2*s + 1))
+    ... 
+    >>> def L(s):
+    ...     return(Kc*G(s))
+    ... 
+    >>> def S(s):
+    ...     return(numpy.linalg.inv((numpy.eye(2) + L(s))))      #SVD of S = 1/(I + L)
+    ... 
+    >>> #utils.perf_Wp(S, 0.05, 0.2, -3, 1)
+    
+    """
+    w = numpy.logspace(wStart, wEnd, 1000)
+    s = w*1j
+    magPlotS1 = numpy.zeros((len(w)))
+    magPlotS3 = numpy.zeros((len(w)))
+    Wpi = numpy.zeros((len(w)))
+    f = 0                                    #f for flag
+    for i in range(len(w)):
+        U, Sv, V = utils.SVD(S(s[i]))
+        magPlotS1[i] = Sv[0]
+        magPlotS3[i] = Sv[-1]
+        if (f < 1 and magPlotS1[i] > 0.707):
+            wB = w[i]
+            f = 1
+    for i in range(len(w)):
+        Wpi[i] = utils.Wp(wB_req, maxSSerror, s[i])                                      
+    plt.figure('MIMO sensitivity S and performance weight Wp')
+    plt.clf()
+    plt.subplot(211)
+    plt.loglog(w, magPlotS1, 'r-', label='Max $\sigma$(S)')
+    plt.loglog(w, 1./Wpi, 'k:', label='|1/W$_P$|', lw=2.)
+    plt.axhline(0.707, color='green', ls=':', lw=2, label='|S| = 0.707')
+    plt.axvline(wB_req, color='blue', ls=':', lw=2)
+    plt.text(wB_req*1.1, 7, 'req wB', color='blue', fontsize=10)
+    plt.axvline(wB, color='green')
+    plt.text(wB*1.1, 0.12, 'wB = %s rad/s'%(numpy.round(wB,3)), color='green', fontsize=10)
+    plt.xlabel('Frequency [rad/s]')
+    plt.ylabel('Magnitude')
+    plt.axis([None, None, 0.1, 10])
+    plt.legend(loc='upper left', fontsize=10, ncol=1)
+    plt.grid(True)
+    plt.subplot(212)
+    plt.semilogx(w, magPlotS1*Wpi, 'r-', label='|W$_P$S|')
+    plt.axhline(1, color='blue', ls=':', lw=2)
+    plt.axvline(wB_req, color='blue', ls=':', lw=2, label='|W$_P$S| = 1')
+    plt.text(wB_req*1.1, numpy.max(magPlotS1*Wpi)*0.95, 'req wB', color='blue', fontsize=10)
+    plt.axvline(wB, color='green')
+    plt.text(wB*1.1, 0.12, 'wB = %s rad/s'%(numpy.round(wB,3)), color='green', fontsize=10)
+    plt.xlabel('Frequency [rad/s]')
+    plt.ylabel('Magnitude')
+    plt.legend(loc='upper right', fontsize=10, ncol=1)
+    fig = plt.gcf()
+    BG = fig.patch
+    BG.set_facecolor('white')
+    plt.grid(True)
+    return(wB)
